pyproject.toml
[project]
name = "test-email"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.2",
    "django>=6.0.3",
    "django-redis>=6.0.0",
    "django-stubs>=5.2.9",
    "factory-boy>=3.3.3",
    "openpyxl>=3.1.5",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.12.0",
    "python-dotenv>=1.2.2",
    "redis>=7.3.0",
    "structlog>=25.5.0",
]
